updated hydra config pathing
